ref: Hoist definitions of `flush`, `close` and `lastEventId` into `@sentry/core`
