+++ packages/node/src/index.ts
@@ -44,11 +44,14 @@ export {
   setTags,
   setUser,
   withScope,
+  lastEventId,
+  flush,
+  close,
 } from '@sentry/core';
 
 export { NodeClient } from './client';
 export { makeNodeTransport } from './transports';
-export { defaultIntegrations, init, defaultStackParser, lastEventId, flush, close, getSentryRelease } from './sdk';
+export { defaultIntegrations, init, defaultStackParser, getSentryRelease } from './sdk';
 export { addRequestDataToEvent, DEFAULT_USER_INCLUDES, extractRequestData } from './requestdata';
 export { deepReadDirSync } from './utils';
